// src/models/user.ts
export class User {
  isUserLoggedIn: boolean;
  id: number;
  name: string;
  facebook_id: string;
  google_id: string;
  email: string;
  first_name: string;
  last_name:string;
  picture: string;
  authToken:string;
  token: string;
  /*constructor(){
    this.email = "";
    this.nombre = "";
    this.apellido = "";
    this.contrasenia = "";
    this.token = "";
  }*/
}